providers: load as a single DB file.
Gathering all providers into a single file has benefits for flash
filesystems on openwrt. Many smaller files consume filesystem idents
and may cause writes of more individual erase sectors than necessary
when installed via the package manager. This likely won't happen if this
app is baked into a sysupgrade, but that is typically an exception to
the normal installation route.

Reading a single file is anyway faster, and compression across a larger
file improves in general vs many smaller files (depending on how the
rootfs is built), although the rootfs is compressed as a single stream
anyway, so this is likely moot.

Removed a few redundant variables (name)

Fixed also Cloudlfare -> Cloudflare (in Tiarap)